+++ packages/cli/bin/utils.js
@@ -6,7 +6,10 @@ const path = require('path');
 const chalk = require('chalk');
 const EventEmitter = require('events').EventEmitter;
 const hasYarn = require('has-yarn');
-const { resolveFileInPackage } = require('@pattern-lab/core/src/lib/resolver');
+const {
+  resolvePackageFolder,
+  resolveFileInPackage,
+} = require('@pattern-lab/core/src/lib/resolver');
 
 /**
  * @name log
@@ -135,7 +138,11 @@ const fetchPackage = packageName =>
     const installCmd = useYarn ? 'add' : 'install';
     try {
       if (packageName) {
-        const cmd = yield spawn(pm, [installCmd, packageName]);
+        const opts = {};
+        if (process.env.projectDir) {
+          opts.cwd = process.env.projectDir;
+        }
+        const cmd = yield spawn(pm, [installCmd, packageName], opts);
         error(cmd.stderr);
       }
     } catch (err) {
@@ -155,7 +162,7 @@ const fetchPackage = packageName =>
 const checkAndInstallPackage = packageName =>
   wrapAsync(function*() {
     try {
-      require.resolve(packageName);
+      resolvePackageFolder(packageName);
       return true;
     } catch (err) {
       debug(