+++ packages/uikit-workshop/src/scripts/components/pl-nav/pl-nav.js
@@ -245,33 +245,33 @@ class Nav extends BaseComponent {
   }
 
   render({ layoutMode }) {
-    const patternTypes = window.navItems.patternTypes;
+    const patternGroups = window.navItems.patternGroups;
 
     return (
       <ol class="pl-c-nav__list pl-js-pattern-nav-target">
-        {patternTypes.map((item, i) => {
+        {patternGroups.map((item, i) => {
           const patternItems = item.patternItems;
 
           return (
-            <NavItem className={`pl-c-nav__item--${item.patternTypeLC}`}>
+            <NavItem className={`pl-c-nav__item--${item.patternGroupLC}`}>
               <NavTitle
-                aria-controls={item.patternTypeLC}
+                aria-controls={item.patternGroupLC}
                 onClick={this.toggleNavPanel}
               >
-                {item.patternTypeUC}
+                {item.patternGroupUC}
               </NavTitle>
               <ol
-                id={item.patternSubtypeUC}
+                id={item.patternSubgroupUC}
                 className={`pl-c-nav__sublist pl-c-nav__sublist--dropdown pl-js-acc-panel`}
               >
-                {item.patternTypeItems.map((patternSubtype, i) => {
+                {item.patternGroupItems.map((patternSubgroup, i) => {
                   return (
                     <NavList
                       elem={this.elem}
-                      category={patternSubtype.patternSubtypeLC}
-                      categoryName={patternSubtype.patternSubtypeUC}
+                      category={patternSubgroup.patternSubgroupLC}
+                      categoryName={patternSubgroup.patternSubgroupUC}
                     >
-                      {patternSubtype.patternSubtypeItems}
+                      {patternSubgroup.patternSubgroupItems}
                     </NavList>
                   );
                 })}