+++ packages/engine-mustache/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.14.0](https://github.com/pattern-lab/patternlab-node/tree/master/packages/engine-mustache/compare/v5.13.3...v5.14.0) (2021-01-12)
+
+**Note:** Version bump only for package @pattern-lab/engine-mustache
+
+
+
+
+
+
 # [5.12.0](https://github.com/pattern-lab/patternlab-node/tree/master/packages/engine-mustache/compare/v5.11.1...v5.12.0) (2020-08-09)
 
 **Note:** Version bump only for package @pattern-lab/engine-mustache